:truck: rename src/utils to src/lib
