+++ src/lib/getStats.ts
