+++ src/lib/pluralize.ts
